poprawa daty rezerwacji,dodanie wyboru opcji cenowej do rezerwacji, automatyczne wylogowywanie uzytkownika w przypadku gdy wygasnie token

+++ src/app/features/my-bookings/my-bookings.component.ts
@@ -13,7 +13,7 @@ import {BookingDetailsComponent} from "../booking-details/booking-details.compon
 })
 export class MyBookingsComponent {
   private reservations: Array<Reservation> = [];
-  displayedColumns: string[] = ['start', 'end', 'menu'];
+  displayedColumns: string[] = ['start', 'end', 'price','menu'];
   dataSource: DataSource<Reservation> = new MatTableDataSource(this.reservations);
 
   constructor(private reservationService: ReservationService,
@@ -36,16 +36,18 @@ export class MyBookingsComponent {
   }
 
   delete(element: Reservation) {
-    this.reservationService.deleteReservation(element.id).subscribe(
-      value => {
-        this.reservationService.getAllReservations().subscribe(
-          value => {
-            this.reservations = value;
-            this.dataSource = new MatTableDataSource(this.reservations);
-          }
-        )
-      }
-    )
+    if(element.id!=null) {
+      this.reservationService.deleteReservation(element.id).subscribe(
+        value => {
+          this.reservationService.getAllReservations().subscribe(
+            value => {
+              this.reservations = value;
+              this.dataSource = new MatTableDataSource(this.reservations);
+            }
+          )
+        }
+      )
+    }
   }
 
   edit(element: Reservation) {
@@ -59,4 +61,35 @@ export class MyBookingsComponent {
       });
     }
 
+  getReservationsPrice(element: Reservation) {
+    if(element.priceEntry != null) {
+      switch (element.priceEntry.price) {
+        case 500:
+          return "Domek MIĘTA- 500zł/noc";
+          break;
+        case 750:
+          return "Domek MIĘTA + strefa SPA- 750zł/noc";
+          break;
+        case 600:
+          return "Domek GRADIENT- 600zł/noc";
+          break;
+        case 850:
+          return "Domek GRADIENT + strefa SPA- 850zł/noc";
+          break;
+        default:
+          return "Błąd";
+      }
+    } else {
+      return 0;
+    }
+  }
+
+  displayDate(date: any) {
+    if(date!=null && date.length>10) {
+      return date.substring(0,10);
+    } else {
+      return "";
+    }
+
+  }
 }
